+++ infra/src/Pos/Infra/Shutdown.hs
@@ -1,9 +1,5 @@
 module Pos.Infra.Shutdown
-       ( module Pos.Infra.Shutdown.Class
-       , module Pos.Infra.Shutdown.Logic
-       , module Pos.Infra.Shutdown.Types
+       ( module X
        ) where
 
-import           Pos.Infra.Shutdown.Class
-import           Pos.Infra.Shutdown.Logic
-import           Pos.Infra.Shutdown.Types
+import           Pos.Sinbin.Shutdown as X
