+++ sinbin/src/Pos/Sinbin/Shutdown/Types.hs
@@ -1,0 +1,17 @@
+{-# LANGUAGE TemplateHaskell #-}
+
+module Pos.Sinbin.Shutdown.Types
+       ( ShutdownContext (..)
+       , shdnIsTriggered
+       ) where
+
+import           Universum
+
+import           Control.Lens (makeLenses)
+
+data ShutdownContext = ShutdownContext
+    { _shdnIsTriggered :: !(TVar Bool)
+    -- ^ If this flag is `True`, then workers should stop.
+    }
+
+makeLenses ''ShutdownContext
